LayoutTests/fast/js/new-array-double-with-holes-expected.txt

Tests that if array allocation profiling causes a new_array to allocate double arrays, then the holes end up being correctly initialized.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS foo([, 1.5], 0) is void 0
PASS successfullyParsed is true

TEST COMPLETE
